achieved compliance with python packaging, local package installation and example works. Reorganized the repo folder structure, so the examples and example_data folders are included into the package. Separated the pyproject.toml file for the package and for the ReadTheDocs build. Included a MANIFEST.in file to include the extra data files.
