+++ MANIFEST.in
@@ -1,0 +1,1 @@
+include tidy_interface/examples/example.csv
