+++ readthedocs-pyproject.toml
@@ -1,0 +1,14 @@
+[build-system]
+requires = ["sphinx", "sphinx-rtd-theme"]
+build-backend = "sphinx.build_api"
+
+[tool.sphinx.conf]
+master_doc = "index"
+
+[tool.sphinx.autodoc]
+members = true
+
+[project]
+name = "tiny_interface"
+authors = [{name = "Tong Zhao", email = "zhao.1991@osu.edu"}]
+dynamic = ["version", "description"]
