+++ tidy_interface/example_data/signals_from_csv.csv
