+++ tidy_interface/signal_manager.py
